requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
requests==2.10.0
ruamel.yaml==0.11.11
suds-jurko==0.6
Werkzeug==0.11.10
